prepare for loading indicator,minor renaming
