+++ next/components/molecules/sections/general/OfficialBoardSection/OfficialBoardSection.tsx
@@ -1,8 +1,8 @@
-import GlobalSearchPageContent from '@components/pages/GlobalSearchPageContent'
+import GlobalSearchSectionContent from '@components/organisms/SearchSection/GlobalSearchSectionContent'
 import React from 'react'
 
 const OfficialBoardSection = () => {
-  return <GlobalSearchPageContent variant="specific" searchOption="officialBoard" />
+  return <GlobalSearchSectionContent variant="specific" searchOption="officialBoard" />
 }
 
 export default OfficialBoardSection
