+++ next/components/molecules/sections/general/RegulationsListSection.tsx
@@ -1,8 +1,8 @@
 import { RegulationsListSectionFragment } from '@backend/graphql'
 import { Typography } from '@bratislava/component-library'
-import SearchBar from '@components/organisms/SearchPage/SearchBar'
-import SearchResults from '@components/organisms/SearchPage/SearchResults'
-import { SearchFilters } from '@components/organisms/SearchPage/useQueryBySearchOption'
+import SearchBar from '@components/organisms/SearchSection/SearchBar'
+import SearchResults from '@components/organisms/SearchSection/SearchResults'
+import { SearchFilters } from '@components/organisms/SearchSection/useQueryBySearchOption'
 import { useTranslations } from 'next-intl'
 import React, { useEffect, useRef, useState } from 'react'
 import { StringParam, useQueryParam, withDefault } from 'use-query-params'
