+++ next/components/molecules/sections/homepage/HomepageTabs/TabPanelOfficialBoard.tsx
@@ -6,8 +6,8 @@ import {
 } from '@backend/ginis/fetchers/officialBoardListFetcher'
 import { Typography } from '@bratislava/component-library'
 import Button from '@components/forms/simple-components/Button'
-import SearchResultCard from '@components/organisms/SearchPage/SearchResultCard'
-import { SearchResult } from '@components/organisms/SearchPage/useQueryBySearchOption'
+import SearchResultCard from '@components/organisms/SearchSection/SearchResultCard'
+import { SearchResult } from '@components/organisms/SearchSection/useQueryBySearchOption'
 import LoadingSpinner from '@components/ui/LoadingSpinner/LoadingSpinner'
 import { useQuery } from '@tanstack/react-query'
 import { base64Encode } from '@utils/base64'
