+++ next/components/organisms/SearchSection/SearchResultCard.tsx
@@ -4,7 +4,7 @@ import { Typography } from '@bratislava/component-library'
 import Icon from '@components/atoms/icon/Icon'
 import ImagePlaceholder from '@components/atoms/ImagePlaceholder'
 import MLink from '@components/forms/simple-components/MLink'
-import { SearchResult } from '@components/organisms/SearchPage/useQueryBySearchOption'
+import { SearchResult } from '@components/organisms/SearchSection/useQueryBySearchOption'
 import { getCategoryColorLocalStyle } from '@utils/colors'
 import { generateImageSizes } from '@utils/generateImageSizes'
 import { isDefined } from '@utils/isDefined'
